+++ packages/time/test/get-running-time.test.js
@@ -1,16 +1,17 @@
+import { rm } from 'node:fs/promises'
 import { join } from 'node:path'
-import { rm } from 'size-limit'
 import { afterEach, expect, it } from 'vitest'
 
 import { getCache, saveCache } from '../cache.js'
 import { cleanCache, getRunningTime } from '../get-running-time.js'
 
 const EXAMPLE = join(__dirname, '../node_modules/nanoid/index.browser.js')
+const CACHE = join(__dirname, '..', '..', '.cache')
 
 afterEach(async () => {
   delete process.env.SIZE_LIMIT_FAKE_TIME
   cleanCache()
-  await rm(join(__dirname, '..', '..', '.cache'))
+  await rm(CACHE, { force: true, recursive: true })
 })
 
 it('calculates running time', async () => {